% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-ts.R, R/build.R
\name{build}
\alias{build}
\alias{build_generator.kg_ts}
\alias{build_generator}
\alias{build_generator.kg_xs}
\title{Compile a time series keras generator meta into fully functional
generator}
\usage{
\method{build_generator}{kg_ts}(x, ...)

build_generator(x, ...)

\method{build_generator}{kg_xs}(x, ...)
}
\arguments{
\item{x}{A time series keras generator meta}

\item{x}{A keras generator meta}
}
\value{
A fully functional time series keras generator object

A fully functional keras generator object
}
\description{
Compile a time series keras generator meta into fully functional
generator

Compile a keras generator meta into fully functional generator
}
