% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-sequences.R
\name{flow_sequences_from_dataframe}
\alias{flow_sequences_from_dataframe}
\title{Setup a sequence data generator specification}
\usage{
flow_sequences_from_dataframe(data, x_names, y_names, lookback, timesteps,
  batch_size = 32, mode = "training")
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame}.}

\item{x_names}{A \code{character} vectors containing
the column names for x variables}

\item{y_names}{A \code{character} vectors containing
the column names for y variables}

\item{lookback}{An \code{integer} specifying lookback period.}

\item{timesteps}{An \code{integer} specifying timesteps length.}

\item{batch_size}{An \code{integer} specifying batch size.}

\item{mode}{A \code{character} vector of length one which specify
the generator behaviour: either \code{"training"} or \code{"prediction"}.
See details for further information.}
}
\value{
A \code{function} which could generate batches of x and/or y arrays
}
\description{
Setup a sequence data generator specification
}
\details{
soon updated
}
\examples{

# import library
library(kerasgenerator)

# create example data
sunspots_df <- as.data.frame(sunspots)

# setup a generator
data_gen <- flow_sequences_from_dataframe(
  data = sunspots_df,
  x_names = "x",
  y_names = "x",
  lookback = 1,
  timesteps = 12,
  batch_size = 32,
  mode = "training"
)

# quick check
arrays <- data_gen()

str(arrays[[1]])
str(arrays[[2]])

}
