% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow-series.R
\name{flow_series_from_dataframe}
\alias{flow_series_from_dataframe}
\alias{tidy_prediction.series_generator}
\title{Setup a time series data generator specification}
\usage{
flow_series_from_dataframe(data, x, y, length_out, stride, lookback,
  timesteps, batch_size = 32, mode = "training")

\method{tidy_prediction}{series_generator}(generator, prediction, ...)
}
\arguments{
\item{data}{A \code{tibble} or \code{data.frame}.}

\item{x}{A \code{character} vectors containing the column names for \code{x} variables}

\item{y}{A \code{character} vectors containing the column names for \code{y} variables}

\item{length_out}{An \code{integer} specifying length out size.}

\item{stride}{An \code{integer} specifying stride size.}

\item{lookback}{An \code{integer} specifying lookback period.}

\item{timesteps}{An \code{integer} specifying timesteps length.}

\item{batch_size}{An \code{integer} specifying batch size.}

\item{mode}{A \code{character} vector of length one which specify
the generator behaviour: either \code{"training"} or \code{"prediction"}.
See details for further information.}

\item{generator}{A reference data generator generated from
\code{kerasgenerator} package.}

\item{prediction}{A \code{matrix} or an \code{array} containing the prediction results}

\item{...}{Further argument for specific generators. See details.}
}
\value{
A \code{function} which could generate batches of x and/or y arrays
}
\description{
Setup a time series data generator specification
}
\details{
soon updated
}
\examples{

# import library
library(kerasgenerator)

# create example data
sunspots_df <- as.data.frame(sunspots)

# setup a generator
data_gen <- flow_series_from_dataframe(
  data = sunspots_df,
  x = "x",
  y = "x",
  length_out = 1,
  stride = 1,
  lookback = 1,
  timesteps = 12,
  batch_size = 32,
  mode = "training"
)

# quick check
arrays <- data_gen()

str(arrays[[1]])
str(arrays[[2]])

}
