% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generator_meta}
\alias{generator_meta}
\title{Get a specific metadata from a generator}
\usage{
generator_meta(x, meta)
}
\arguments{
\item{x}{A generator generated using \code{kerasgenerator}.}

\item{meta}{A \code{character} vectors of length one to specify which
metadata to get. \strong{Optionally}, the function also support \code{"all"} value
to return all metadata. See details.}
}
\value{
An object with class according to specified metadata
}
\description{
Get a specific metadata from a generator
}
\details{
soon updated
}
\examples{

# import library
library(kerasgenerator)

# create example data
sunspots_df <- as.data.frame(sunspots)

# setup a generator
data_gen <- flow_series_from_dataframe(
  data = sunspots_df,
  x = "x",
  y = "x",
  length_out = 1,
  stride = 1,
  lookback = 1,
  timesteps = 12,
  batch_size = 32,
  mode = "training"
)

# check generator mode
generator_meta(data_gen, "mode")

# get all metadata as list
meta <- generator_meta(data_gen, "all")

meta[["batch_size"]]

}
