% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{generator_mode}
\alias{generator_mode}
\title{Set/modify a generator's mode}
\usage{
generator_mode(x, new_mode)
}
\arguments{
\item{x}{A generator generated using \code{kerasgenerator}.}

\item{new_mode}{new mode that would be assigned to the generator.}
}
\value{
An modified data generator (copied as new environment)
}
\description{
Set/modify a generator's mode
}
\details{
soon updated
}
\examples{

# import library
library(kerasgenerator)

# create example data
sunspots_df <- as.data.frame(sunspots)

# setup a generator
data_gen <- flow_series_from_dataframe(
  data = sunspots_df,
  x = "x",
  y = "x",
  length_out = 1,
  stride = 1,
  lookback = 1,
  timesteps = 12,
  batch_size = 32,
  mode = "training"
)

# check generator mode
generator_meta(data_gen, "mode")

# modify generator mode
data_gen_modified <- generator_mode(data_gen, "prediction")

# compare old and new generators
generator_meta(data_gen, "mode")
generator_meta(data_gen_modified, "mode")

}
