% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_batch}
\alias{get_batch}
\title{Get a sample batch from a generator}
\usage{
get_batch(gen, n = 1)
}
\arguments{
\item{n}{The number of specific batch partition to get. Default to the
first partition}

\item{x}{A keras generator object}
}
\value{
A batch of keras array(s)
}
\description{
A helper function to get a sample batch in the form of keras
array(s)
}
