% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_meta}
\alias{get_meta}
\title{Get a meta data from a generator}
\usage{
get_meta(gen, meta)
}
\arguments{
\item{meta}{A specific meta data inside the generator}

\item{x}{A keras generator object}
}
\value{
The specified meta data object
}
\description{
A helper function to get a specific meta data from generator
environment
}
