% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\name{inject_recipe}
\alias{inject_recipe}
\title{Inject \code{recipes} object for preprocessing on the fly}
\usage{
inject_recipe(x, rec)
}
\arguments{
\item{x}{A keras generator object}

\item{rec}{A prepared \code{recipes} object}
}
\value{
An updated keras generator object
}
\description{
Inject \code{recipes} object for preprocessing on the fly
}
\concept{injection}
