% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{filter}
\alias{group_by}
\alias{mutate}
\alias{select}
\alias{slice}
\alias{summarise}
\alias{ungroup}
\alias{gather}
\alias{spread}
\alias{glimpse}
\alias{str_replace_all}
\alias{head}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{ungroup}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}}

  \item{stringr}{\code{\link[stringr]{str_replace_all}}}

  \item{tibble}{\code{\link[tibble]{glimpse}}}

  \item{tidyr}{\code{\link[tidyr]{gather}}, \code{\link[tidyr]{spread}}}

  \item{utils}{\code{\link[utils]{head}}}
}}

