% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_generator}
\alias{reset_generator}
\title{Reset a generator into a specific partition}
\usage{
reset_generator(gen, to = 1)
}
\arguments{
\item{to}{The number of specific batch partition to be set. Default to the
first partition}

\item{x}{A keras generator object}
}
\value{
An updated keras generator object
}
\description{
A helper function to reset the generator current partition
}
