% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip.R
\name{select_x_y}
\alias{select_x_y}
\alias{select_x}
\alias{select_y}
\title{Select feature and target for x and y arrays}
\usage{
select_x(x, ...)

select_y(x, ...)
}
\arguments{
\item{x}{A keras generator object}

\item{...}{Variable(s) of selection}
}
\value{
An updated keras generator object
}
\description{
Select variables for x and y arrays with \code{dplyr}-like API
}
\concept{manipulation}
