% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_generator.R
\name{series_generator}
\alias{series_generator}
\title{Time series data generator for training and testing}
\usage{
series_generator(data, x, y, lookback, timesteps, start_index, end_index,
  batch_size, return_target = TRUE)
}
\arguments{
\item{data}{A data frame or matrix containing the time series data.}

\item{x}{Numeric or character vectors that identify
which column(s) in the \code{data} is/are the feature(s).}

\item{y}{Numeric or character vectors that identify
which column(s) in the \code{data} is/are the target(s).}

\item{lookback}{A numeric vector which identify
the number of lookback period.}

\item{timesteps}{A numeric vector which identify
the number of timesteps length.}

\item{start_index}{A numeric vector which identify
the first row index of the the \code{data}
that would be included as a sample.}

\item{end_index}{A numeric vector which identify
the first row index of the the \code{data}
that would be included as a sample.}

\item{batch_size}{A numeric vector which identify
how many observation is considered as a batch.}

\item{return_target}{A boolean indicating if the generator.
should return the target or not.}
}
\value{
If \code{return_target} is set to \code{TRUE},
the generator will returning and object of \code{list} containing
a batch of \code{x} and \code{y arrays}.
}
\description{
A data generator function for training and testing
a supervised time series Keras model
}
\seealso{
Other time series generators: \code{\link{forecast_generator}}
}
\concept{time series generators}
