% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_generator}
\alias{set_generator}
\title{Set-up a cross-sectional data generator meta}
\usage{
set_generator(data, batch_size = 32, shuffle = TRUE, seed = 1)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl} containing the cross-sectional data}

\item{batch_size}{An \code{integer} value indicating how many observation is
considered as a batch}

\item{shuffle}{A \code{logical} indicating whether to shuffle the observation
within the batch}

\item{seed}{An optional \code{integer} value for passing to \code{set.seed()} if
\code{shuffle} is \code{TRUE}}
}
\value{
A keras generator meta
}
\description{
An initial function to define a cross-sectional data generator
}
\seealso{
Other cross-section generator: \code{\link{set_series_generator}}
}
\concept{cross-section generator}
