% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-prediction.R
\name{tidy_prediction}
\alias{tidy_prediction}
\title{Helper function for tidying \code{keras}' prediction results}
\usage{
tidy_prediction(generator, prediction, ...)
}
\arguments{
\item{generator}{A reference data generator generated from
\code{kerasgenerator} package.}

\item{prediction}{A \code{matrix} or an \code{array} containing the prediction results}

\item{...}{Further argument for specific generators. See details.}
}
\value{
A \code{tible} containing tidied prediction results
}
\description{
Helper function for tidying \code{keras}' prediction results
}
\details{
soon updated
}
