% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-xs.R, R/build.R
\name{build_generator.kg_xs}
\alias{build_generator.kg_xs}
\alias{build}
\alias{build_generator}
\title{Compile a keras generator meta into fully functional generator}
\usage{
\method{build_generator}{kg_xs}(x, ...)

build_generator(x, ...)
}
\arguments{
\item{x}{A keras generator meta}
}
\value{
A fully functional keras generator object
}
\description{
Compile a keras generator meta into fully functional generator
}
