% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_generator.R
\name{forecast_generator}
\alias{forecast_generator}
\title{Data generator for forecasting using time series Keras model}
\usage{
forecast_generator(data, x, timesteps = 1, start_index = NULL,
  end_index = NULL, batch_size = 32, prep_funs = NULL)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} containing
the time series data.}

\item{x}{\code{numeric} or \code{character} vectors that identify
which column(s) in the \code{data} is/are the feature(s).}

\item{timesteps}{A \code{numeric} vector which identify
the number of timesteps length.}

\item{start_index}{A \code{numeric} vector which identify
the first row index of the the \code{data}
that would be included as a sample.}

\item{end_index}{A \code{numeric} vector which identify
the first row index of the the \code{data}
that would be included as a sample.}

\item{batch_size}{A \code{numeric} vector which identify
how many observation is considered as a batch.}

\item{prep_funs}{A custom function for preprocess the \code{data}
on the fly, optional.}
}
\value{
The generator will returning an object of \code{list} containing
a batch of \code{x_array}.
}
\description{
A data generator function for forecasting using
a supervised time series Keras model
}
\seealso{
Other time series generators: \code{\link{series_generator}}
}
\concept{time series generators}
