% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_generator.R
\name{forecast_generator}
\alias{forecast_generator}
\title{Time series data generator for forecasting}
\usage{
forecast_generator(data, x, timesteps, start_index = NULL,
  end_index = NULL, batch_size)
}
\arguments{
\item{data}{A data frame containing the time series data.}

\item{x}{Numeric or character vectors that identify
which column(s) in the \code{data} is/are the feature(s).}

\item{timesteps}{A numeric vector which identify
the number of timesteps length.}

\item{start_index}{A numeric vector which identify
the first row index of the the \code{data}
that would be included as a sample.}

\item{end_index}{A numeric vector which identify
the first row index of the the \code{data}
that would be included as a sample.}

\item{batch_size}{A numeric vector which identify
how many observation is considered as a batch.}
}
\value{
The generator will returning and object of \code{list} containing
a batch of \code{x} arrays.
}
\description{
A data generator function for forecasting
a supervised time series Keras model
}
\seealso{
Other time series generators: \code{\link{series_generator}}
}
\concept{time series generators}
