% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_series_generator}
\alias{set_series_generator}
\title{Set-up a time series data generator meta}
\usage{
set_series_generator(data, lookback = 1, timesteps = 1,
  batch_size = 32, shuffle = FALSE, seed = 1)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tbl}, or \code{tbl_ts} containing
the time series data}

\item{lookback}{A \code{numeric} vector which identify the number
of lookback period}

\item{timesteps}{A \code{numeric} vector which identify the number
of timesteps length}

\item{batch_size}{An \code{integer} value indicating how many observation is
considered as a batch}

\item{shuffle}{A \code{logical} indicating whether to shuffle the observation
within the batch}

\item{seed}{An optional \code{integer} value for passing to \code{set.seed()} if
\code{shuffle} is \code{TRUE}}

\item{timesteps}{Variable name representing the time index}
}
\value{
A keras generator meta
}
\description{
An initial function to define a cross-sectional data generator
}
\seealso{
Other cross-section generator: \code{\link{set_generator}}
}
\concept{cross-section generator}
